Move 'formatBytes' function to helpers file
