+++ Source/SelfService/Web/microservice/microserviceDetails/configurationFilesSection/filesSection/validateFileDialog.tsx
@@ -7,19 +7,9 @@ import { ConfirmDialog } from '@dolittle/design-system';
 
 import { Divider, List, ListItem, Typography } from '@mui/material';
 
+import { formatBytes } from '../../../helpers/';
+
 const MAX_CONFIGMAP_ENTRY_SIZE = 3145728;
-
-function formatBytes(bytes: number, decimals = 2) {
-    if (!+bytes) return '0 Bytes';
-
-    const k = 1024;
-    const dm = decimals < 0 ? 0 : decimals;
-    const sizes = ['Bytes', 'KB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB'];
-
-    const i = Math.floor(Math.log(bytes) / Math.log(k));
-
-    return `${parseFloat((bytes / Math.pow(k, i)).toFixed(dm))} ${sizes[i]}`;
-};
 
 type ValidateFileDialogProps = {
     invalid: {
